+++ modules/FiAKEr/iAMeasureSelectionDlg.cpp
@@ -126,7 +126,7 @@ int iAMeasureSelectionDlg::optimizeMeasureIdx() const
 		if (m_model->item(row, 3)->checkState() == Qt::Checked)
 		{
 			auto m = measures();
-			for (int mIdx = 0; mIdx < m.size(); ++mIdx)
+			for (size_t mIdx = 0; mIdx < m.size(); ++mIdx)
 			{
 				if (m[mIdx].first == row)
 				{
